+++ autotransform/scripts/main.py
@@ -14,7 +14,7 @@ Different CLI commands are handled as subparsers. Includes the following command
 
 from argparse import ArgumentParser
 
-from autotransform.scripts.commands import manage, run, schedule, update
+from autotransform.scripts.commands import config, manage, run, schedule, update
 
 
 def get_arg_parser() -> ArgumentParser:
@@ -61,6 +61,14 @@ def get_arg_parser() -> ArgumentParser:
         aliases=["u"],
     )
     update.add_args(update_parser)
+
+    # Add config command
+    config_parser = subparsers.add_parser(
+        "config",
+        help="Update/view AutoTransform configs",
+        aliases=["c"],
+    )
+    config.add_args(config_parser)
 
     return parser